moving rt commands in plugin.go
